move sanitize-path and ensure-trailing-slash to core util
