+++ ui/lib/core/app/utils/sanitize-path.js
@@ -1,0 +1,1 @@
+export { ensureTrailingSlash, sanitizePath } from 'core/utils/sanitize-path';
